requirements.txt
# Pipeline de Transcription CISR - Dépendances Python
# Installation: pip install -r requirements.txt

# Documents Word
python-docx>=1.1.0
docxcompose>=1.4.0

# Excel
openpyxl>=3.1.0

# API REST
requests>=2.31.0

# Variables d'environnement
python-dotenv>=1.0.0

# Tests
pytest>=8.0.0
pytest-cov>=4.1.0

# Audio (optionnel - pour analyse préliminaire)
# pydub>=0.25.1
